requirements.txt
tk
Pillow
requests
ollama